# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlloyLab API",
    "description": "Autonomous computational materials-science platform (V0: Ising critical-region scientist).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/benchmarks": {
      "get": {
        "tags": [
          "benchmarks"
        ],
        "summary": "List Benchmarks",
        "operationId": "list_benchmarks_benchmarks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BenchmarkRead"
                  },
                  "type": "array",
                  "title": "Response List Benchmarks Benchmarks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "benchmarks"
        ],
        "summary": "Create Benchmark",
        "operationId": "create_benchmark_benchmarks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/benchmarks/{benchmark_id}": {
      "get": {
        "tags": [
          "benchmarks"
        ],
        "summary": "Get Benchmark",
        "operationId": "get_benchmark_benchmarks__benchmark_id__get",
        "parameters": [
          {
            "name": "benchmark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Benchmark Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenchmarkCreate": {
        "properties": {
          "problem": {
            "$ref": "#/components/schemas/BenchmarkProblem",
            "default": "ising"
          },
          "strategies": {
            "items": {
              "$ref": "#/components/schemas/StrategyName"
            },
            "type": "array",
            "title": "Strategies",
            "default": [
              "random",
              "grid",
              "uncertainty"
            ]
          },
          "budget": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 4.0,
            "title": "Budget",
            "default": 12
          },
          "seeds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Seeds",
            "default": [
              1,
              2,
              3
            ]
          },
          "lattice_size": {
            "type": "integer",
            "maximum": 48.0,
            "minimum": 8.0,
            "title": "Lattice Size",
            "default": 16
          },
          "temperature_min": {
            "type": "number",
            "title": "Temperature Min",
            "default": 1.5
          },
          "temperature_max": {
            "type": "number",
            "title": "Temperature Max",
            "default": 3.5
          }
        },
        "type": "object",
        "title": "BenchmarkCreate"
      },
      "BenchmarkProblem": {
        "type": "string",
        "enum": [
          "ising",
          "alloy"
        ],
        "title": "BenchmarkProblem"
      },
      "BenchmarkRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Results"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "config",
          "results",
          "summary",
          "error",
          "created_at",
          "completed_at"
        ],
        "title": "BenchmarkRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StrategyName": {
        "type": "string",
        "enum": [
          "agent",
          "random",
          "grid",
          "uncertainty"
        ],
        "title": "StrategyName"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}